requirements.txt
langchain
langchain-community
langchain-mongodb
langchain-groq
langchain-huggingface
pymongo[srv]
sentence-transformers
pypdf
python-dotenv
streamlit
tiktoken
torch
torchvision
